Precompile regex and move file content function into input.rs where it belongs
