--- .clippy.toml
@@ -1,1 +1,0 @@
-allowed-duplicate-crates = ["syn"]
