+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "textgridde-rs"
-version = "0.1.2"
+version = "0.1.4"
 edition = "2021"
 description = "A library for dealing with Praat TextGrid files. MIT licensed."
 authors = ["Caleb Long <cmlong02@outlook.com>"]
